+++ CHANGELOG.md
@@ -26,7 +26,7 @@ Change categories:
 5. pdr: Add pldm_pdr_get_terminus_handle() API
 6. pdr: Add related decode_entity_auxiliary_names_pdr() APIs
 7. fw_update: Add encode req & decode resp for get_downstream_fw_params
-8. platform: Add decode_pldm_platform_cper_event_data() API
+8. platform: Add decode_pldm_platform_cper_event() API
 
 ### Changed